requirements.txt
Flask==3.0.0
pymongo[srv]==4.6.1
Pillow==10.1.0
numpy==1.26.2
tqdm==4.66.1
transformers==4.36.0
sentencepiece==0.1.99
ftfy==6.1.3
regex==2023.12.25
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1